moved the db-stub into a stubs folder
